creating charge, creating subscription, creating products, creating paymentMethod, creating cards. Exception handler

+++ src/main/java/com/inikitagricenko/demo/stripe/entity/ProductEntity.java
@@ -33,7 +33,7 @@ public class ProductEntity {
 	private String description;
 
 	@Column(name = "price")
-	private Long price;
+	private float price;
 
 	@Column(name = "quantity")
 	private Long quantity;
